refactor: move PostHog event scrubber into utils
Keep PostHogProvider focused on init; scrub helpers live in
scrub-posthog-event.ts.
